+++ web/admin/src/App.vue
@@ -2,12 +2,14 @@
   <n-config-provider :theme="null">
     <n-message-provider>
       <n-notification-provider>
-        <router-view />
+        <n-dialog-provider>
+          <router-view />
+        </n-dialog-provider>
       </n-notification-provider>
     </n-message-provider>
   </n-config-provider>
 </template>
 
 <script setup lang="ts">
-import { NConfigProvider, NMessageProvider, NNotificationProvider } from 'naive-ui';
+import { NConfigProvider, NMessageProvider, NNotificationProvider, NDialogProvider } from 'naive-ui';
 </script>